requirements.txt
bottle==0.12.19
certifi==2020.11.8
chardet==3.0.4
httplib2==0.18.1
idna==2.10
Pillow==8.0.1
psutil==5.7.3
requests==2.25.0
urllib3==1.26.2
yadisk==1.2.14
